feat: add categoria CRUD

+++ controller/ProdutosController.js
@@ -22,7 +22,7 @@ exports.cadastrarProduto = async function (req, res, next) {
                 nome: fields.nome[0],
                 preco: fields.preco[0],
                 imagem: `${files.imagem[0].newFilename}.${aux2[1]}`,
-                categoria_id: fields.categoria_id[0]
+                categoria_id: fields.categoria[0]
             })
             res.send(JSON.stringify({ status: 'sucesso', mensagem: aux }))
         } else {
@@ -30,7 +30,7 @@ exports.cadastrarProduto = async function (req, res, next) {
                 nome: fields.nome[0],
                 preco: fields.preco[0],
                 imagem: null,
-                categoria_id: fields.categoria_id[0]
+                categoria_id: fields.categoria[0]
             })
             res.send(JSON.stringify({ status: 'sucesso', mensagem: aux }))
         }
@@ -59,6 +59,35 @@ exports.listarCategorias = async (req, res, next) => {
     try {
         const resultado = await Categoria.findAll()
         res.send(JSON.stringify({ status: "sucesso", mensagem: resultado }))
+    }
+    catch (erro) {
+        console.log(erro)
+        res.send(JSON.stringify({ status: 'falha', mensagem: 'ocorreu um erro ao consumir a api' }))
+    }
+}
+
+exports.criarCategoria = async (req, res, next) => {
+    try {
+        const resultado = await Categoria.create({
+            categoria: req.body.categoria
+        })
+        res.send(JSON.stringify({ status: "sucesso", mensagem: resultado }))
+    }
+    catch (erro) {
+        console.log(erro)
+        res.send(JSON.stringify({ status: 'falha', mensagem: 'ocorreu um erro ao consumir a api' }))
+    }
+}
+
+exports.deletarCategoria = async (req, res, next) => {
+    try {
+        if (req.params.id) {
+            const resultado = await Categoria.destroy({ where: { id: req.params.id } })
+            res.send(JSON.stringify({ status: "sucesso", mensagem: resultado }))
+        } else {
+            throw new Error("Falha ao deletar, faltou id.")
+        }
+
     }
     catch (erro) {
         console.log(erro)
@@ -99,11 +128,13 @@ exports.alterarFoto = async (req, res, next) => {
                 nomeFoto = `${files.imagem[0].newFilename}` + `.${aux2[1]}`
                 const novaFoto = await Produto.update({ imagem: nomeFoto }, { where: { id: consultaFotoAnterior.dataValues.id } })
 
-                caminhoFotoAntiga = path.join(__dirname + `./../public/img/${consultaFotoAnterior.dataValues.imagem}`)
-                excluirFoto = fs.unlink(caminhoFotoAntiga, (erro) => {
-                    if (erro) throw erro;
-                    console.log('Foto antiga excluida com sucesso')
-                })
+                if (consultaFotoAnterior.dataValues.imagem != null) {
+                    caminhoFotoAntiga = path.join(__dirname + `./../public/img/${consultaFotoAnterior.dataValues.imagem}`)
+                    excluirFoto = fs.unlink(caminhoFotoAntiga, (erro) => {
+                        if (erro) throw erro;
+                        console.log('Foto antiga excluida com sucesso')
+                    })
+                }
                 res.send(JSON.stringify({ status: "sucesso", mensagem: "a foto do produto foi alterada com sucesso" }))
             } else {
                 res.send(JSON.stringify({ status: "falha", mensagem: "erro ao localizar o produto" }))
@@ -121,11 +152,13 @@ exports.deletarProduto = async (req, res, next) => {
     try {
         const consultaFotoAnterior = await Produto.findOne({ where: { id: req.body.id } })
         const resultado = await Produto.destroy({ where: { id: req.body.id } })
-        caminhoFotoAntiga = path.join(__dirname + `./../public/img/${consultaFotoAnterior.dataValues.imagem}`)
-        excluirFoto = fs.unlink(caminhoFotoAntiga, (erro) => {
-            if (erro) throw erro;
-            console.log('Foto antiga excluida com sucesso')
-        })
+        if (consultaFotoAnterior.dataValues.imagem != null) {
+            caminhoFotoAntiga = path.join(__dirname + `./../public/img/${consultaFotoAnterior.dataValues.imagem}`)
+            excluirFoto = fs.unlink(caminhoFotoAntiga, (erro) => {
+                if (erro) throw erro;
+                console.log('Foto antiga excluida com sucesso')
+            })
+        }
         res.send(JSON.stringify({ status: "sucesso", mensagem: resultado }))
     } catch (erro) {
         res.send(JSON.stringify({ status: 'falha', mensagem: 'ocorreu um erro ao consumir a api' }))
